Modelo/llm/rag/scripts/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
tqdm>=4.65.0
numpy>=1.24.0
